// frontend-angular/src/app/core/generated-api/model/rolResponse.ts
/**
 * API - Sistema de Citas Medicas
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 * Un rol del sistema (tabla Roles)
 */
export interface RolResponse { 
    readonly id: number;
    nombre: string;
}