// frontend-angular/src/app/core/generated-api/model/usuarioResponse.ts
/**
 * API - Sistema de Citas Medicas
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 * Representacion de una cuenta de usuario del sistema
 */
export interface UsuarioResponse { 
    readonly id: number;
    nombreCompleto: string;
    username: string;
    /**
     * Nombre del rol asignado a esta cuenta
     */
    rolNombre: string;
    activo: boolean;
    /**
     * Identificador del registro de Medico vinculado a esta cuenta, presente unicamente cuando el rol asignado es Medico
     */
    medicoId?: number | null;
}